인스타 로그인 회원가입

+++ instagram-clone-main/instagram-clone-main/src/main/java/kr/or/spring/instagram_clone/dto/UserRole.java
@@ -1,31 +1,31 @@
 package kr.or.spring.instagram_clone.dto;
 
 public class UserRole {
-	private Long id;
-	private Long userId;
+	private int id;
+	private int userId;
 	private String roleName;
 
 	public UserRole() {
 	}
 
-	public UserRole(Long userId, String roleName) {
+	public UserRole(int userId, String roleName) {
 		this.userId = userId;
 		this.roleName= roleName;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 
-	public Long getUserId() {
+	public int getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Long userId) {
+	public void setUserId(int userId) {
 		this.userId = userId;
 	}
